feat(markdown): shared HorizontalScrollView + smarter wheel behavior
Extract useHorizontalWheelScroll into a reusable HorizontalScrollView
component shared by code blocks and tables. Vertical mouse wheel no
longer hijacks the page scroll when hovering over a scrollable block
unless Shift is held; touchpad horizontal swipe scrolls natively.

Also cap table column maxWidth at 360 so long cells wrap instead of
forcing excessive horizontal scroll.
