+++ packages/happy-app/sources/components/HorizontalScrollView.tsx
@@ -1,0 +1,47 @@
+import * as React from 'react';
+import { Platform, ScrollView, ScrollViewProps } from 'react-native';
+
+// Horizontal-dominant wheel (touchpad two-finger swipe) scrolls natively.
+// Vertical wheel only hijacks to horizontal when Shift is held, so pointer-over
+// the block doesn't hostage vertical page scroll.
+function useHorizontalWheelScroll() {
+    const ref = React.useRef<ScrollView>(null);
+    React.useEffect(() => {
+        if (Platform.OS !== 'web' || !ref.current) return;
+        const node = (ref.current as any)?.getScrollableNode?.() ?? (ref.current as any);
+        if (!node || !node.addEventListener) return;
+        const handler = (e: WheelEvent) => {
+            const el = node as HTMLElement;
+            const maxScroll = el.scrollWidth - el.clientWidth;
+            if (maxScroll <= 0) return;
+            if (Math.abs(e.deltaX) >= Math.abs(e.deltaY)) return;
+            if (!e.shiftKey) return;
+            e.preventDefault();
+            e.stopPropagation();
+            el.scrollLeft += e.deltaY;
+        };
+        node.addEventListener('wheel', handler, { passive: false });
+        return () => node.removeEventListener('wheel', handler);
+    });
+    return ref;
+}
+
+type Props = Omit<ScrollViewProps, 'horizontal'>;
+
+export function HorizontalScrollView(props: Props) {
+    const {
+        showsHorizontalScrollIndicator = true,
+        nestedScrollEnabled = true,
+        ...rest
+    } = props;
+    const ref = useHorizontalWheelScroll();
+    return (
+        <ScrollView
+            ref={ref}
+            horizontal
+            showsHorizontalScrollIndicator={showsHorizontalScrollIndicator}
+            nestedScrollEnabled={nestedScrollEnabled}
+            {...rest}
+        />
+    );
+}
